services/api_gateway/requirements.txt
fastapi
uvicorn[standard]==0.34.0
websockets==14.0
wsproto==1.2.0
pydantic
python-dotenv
httpx
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
google-auth
requests
websockets
